refactor video options to bottom sheet
