+++ app/src/main/java/com/github/libretube/adapters/TrendingAdapter.kt
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.FragmentManager
 import androidx.recyclerview.widget.RecyclerView
 import com.github.libretube.databinding.TrendingRowBinding
-import com.github.libretube.dialogs.VideoOptionsDialog
+import com.github.libretube.sheets.VideoOptionsBottomSheet
 import com.github.libretube.extensions.formatShort
 import com.github.libretube.extensions.setFormattedDuration
 import com.github.libretube.extensions.setWatchProgressLength
@@ -63,8 +63,8 @@ class TrendingAdapter(
             }
             val videoId = trending.url!!.toID()
             root.setOnLongClickListener {
-                VideoOptionsDialog(videoId)
-                    .show(childFragmentManager, VideoOptionsDialog::class.java.name)
+                VideoOptionsBottomSheet(videoId)
+                    .show(childFragmentManager, VideoOptionsBottomSheet::class.java.name)
                 true
             }
             watchProgress.setWatchProgressLength(videoId, trending.duration!!)
